requirements.txt
Django==1.7.1
djangorestframework==3.0.2
fudge==1.0.3
google-api-python-client==1.3.1
httplib2==0.9
isodate==0.5.0
oauth2==1.5.211
oauth2client==1.4.4
psycopg2==2.5.4
pyasn1==0.1.7
pyasn1-modules==0.0.5
requests==2.4.1
rsa==3.1.4
simplejson==3.6.5
six==1.8.0
soundcloud==0.4.1
uritemplate==0.6
